Final version of hw3

+++ src/main/scala/ru/ifmo/ctddev/ml/hw3/Mail.scala
@@ -4,7 +4,7 @@ import java.io.InputStream
 import java.util.Scanner
 
 
-case class Mail(subject: Seq[Int], body: Seq[Int])
+case class Mail(subject: Seq[Long], body: Seq[Long])
 
 case class MailWithVerdict(mail: Mail, verdict: Boolean)
 
@@ -14,6 +14,6 @@ object Mail {
     val subjectLine = scanner.nextLine().substring("Subject: ".length)
     scanner.nextLine()
     val bodyLine = scanner.nextLine()
-    Mail(subjectLine.split(' ').filterNot(_.isEmpty).map(_.toInt), bodyLine.split(' ').filterNot(_.isEmpty).map(_.toInt))
+    Mail(subjectLine.split(' ').filterNot(_.isEmpty).map(_.toLong), bodyLine.split(' ').filterNot(_.isEmpty).map(_.toLong))
   }
 }
